==> Hazime-zayour_Maha_studentA/Code/1.scraping_renewable_energy_data.ipynb <==
import mechanicalsoup
import pandas as pd
import numpy as np
from io import StringIO

# Create a browser object
browser = mechanicalsoup.Browser()

# URL of the page to scrape
url = 'https://de.wikipedia.org/wiki/Erneuerbare_Energien_in_Europa'

# Load the webpage
page = browser.get(url)

# Parse the HTML and find the table
table = page.soup.find_all('table')[1]

# Use pandas to read the HTML table
html_table = str(table)

# Replace ',' with '.' in the HTML table to fix the missing decimal points
html_table = html_table.replace(',', '.')

df = pd.read_html(StringIO(html_table), header=0)[0]

# Specify the column indices you want to keep (0, 2, 3)
desired_column_indices = [0, 2, 3]

# Select only the desired columns
df = df.iloc[:, desired_column_indices]

# Display the first few rows of the DataFrame
print(df.head())

# Save the DataFrame to a CSV file
csv_file = '1.renewable_energy_stage1.csv'
df.to_csv(csv_file, index=False)

# Print a message confirming the file save
print(f'Table saved to {csv_file}')
