==> Hazime-zayour_Maha_studentA/Code/4.merged_data.ipynb <==
# Merging Data 
#This Jupyter Notebook focuses on Merging the scraped datasets the  Air Quality dataset and 
# the Renewable Energy dataset. Then Pivot and restructure the Dataset.


import pandas as pd

# Load the first CSV file (df1)
file1 = '3.cleaned_airquality_data_stage3.csv'
df1 = pd.read_csv(file1)

# Load the second CSV file (df2)
file2 = '3.cleaned_renewable_energy_stage3.csv'
df2 = pd.read_csv(file2)

# Ensure 'Year' is the same data type in both data frames
df1['Year'] = df1['Year'].astype(int)
df2['Year'] = df2['Year'].astype(int)

# Merge df2 into df1 based on 'Country' and 'Year'
merged_df = df1.merge(df2, on=['Country', 'Year'], how='left')

# Display the merged DataFrame
print(merged_df.head())



# Save the merged DataFrame to a new CSV file
merged_df.to_csv('AirQuality_RenewableEngergy_data.csv', index=False)
